+++ example/lib/serializers.g.dart
@@ -8,68 +8,15 @@ part of serializers;
 
 Serializers _$serializers = (new Serializers().toBuilder()
       ..add(Account.serializer)
-      ..add(BoundGenericValue.serializer)
       ..add(Cat.serializer)
-      ..add(CollectionGenericValue.serializer)
-      ..add(Collections.serializer)
       ..add(CompoundValue.serializer)
-      ..add(EnumWithInt.serializer)
       ..add(Fish.serializer)
-      ..add(GenericContainer.serializer)
       ..add(GenericValue.serializer)
       ..add(SimpleValue.serializer)
-      ..add(TestEnum.serializer)
       ..add(ValidatedValue.serializer)
-      ..add(ValueWithInt.serializer)
       ..add(VerySimpleValue.serializer)
-      ..addBuilderFactory(
-          const FullType(BuiltList, const [const FullType(int)]),
-          () => new ListBuilder<int>())
-      ..addBuilderFactory(
-          const FullType(BuiltSet, const [const FullType(String)]),
-          () => new SetBuilder<String>())
-      ..addBuilderFactory(
-          const FullType(
-              BuiltMap, const [const FullType(String), const FullType(int)]),
-          () => new MapBuilder<String, int>())
-      ..addBuilderFactory(
-          const FullType(BuiltListMultimap,
-              const [const FullType(int), const FullType(bool)]),
-          () => new ListMultimapBuilder<int, bool>())
-      ..addBuilderFactory(
-          const FullType(BuiltSetMultimap,
-              const [const FullType(String), const FullType(bool)]),
-          () => new SetMultimapBuilder<String, bool>())
-      ..addBuilderFactory(
-          const FullType(BuiltList, const [const FullType(int)]),
-          () => new ListBuilder<int>())
-      ..addBuilderFactory(
-          const FullType(BuiltSet, const [const FullType(String)]),
-          () => new SetBuilder<String>())
-      ..addBuilderFactory(
-          const FullType(
-              BuiltMap, const [const FullType(String), const FullType(int)]),
-          () => new MapBuilder<String, int>())
-      ..addBuilderFactory(
-          const FullType(BuiltListMultimap,
-              const [const FullType(int), const FullType(bool)]),
-          () => new ListMultimapBuilder<int, bool>())
-      ..addBuilderFactory(
-          const FullType(BuiltSetMultimap,
-              const [const FullType(String), const FullType(bool)]),
-          () => new SetMultimapBuilder<String, bool>())
       ..addBuilderFactory(
           const FullType(BuiltMap,
               const [const FullType(String), const FullType(JsonObject)]),
-          () => new MapBuilder<String, JsonObject>())
-      ..addBuilderFactory(
-          const FullType(GenericValue, const [const FullType(String)]),
-          () => new GenericValueBuilder<String>())
-      ..addBuilderFactory(
-          const FullType(BoundGenericValue, const [const FullType(double)]),
-          () => new BoundGenericValueBuilder<double>())
-      ..addBuilderFactory(
-          const FullType(
-              CollectionGenericValue, const [const FullType(String)]),
-          () => new CollectionGenericValueBuilder<String>()))
+          () => new MapBuilder<String, JsonObject>()))
     .build();